[fix] Removed CSV files if they exist
If there are any CSV files in the data directory, this commit removes them. If no CSV files are found, it prints a message saying "No CSV files to remove."

+++ start.sh
@@ -2,7 +2,11 @@
 
 printf '\033c'; clear;
 
-rm ./data/*.csv;
+if ls ./data/*.csv 1> /dev/null 2>&1; then
+    rm ./data/*.csv
+else
+    echo "No CSV files to remove."
+fi
 
 #  brew install pipx
 #  pipx install poetry
